refactor(type4): split axis source templates
